requirements.txt
telethon==1.36.0
ccxt==4.4.93
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
tenacity==9.0.0

# Web UI
fastapi==0.115.4
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
jinja2==3.1.4